requirements.txt
claude-code-sdk>=0.0.25
python-dotenv>=1.0.0
google-genai>=1.0.0
